Ubuntu/Linux destegi, VLM'in vLLM'e tasinmasi ve FPS optimizasyonlari
ORTAM
- Sistemi_Baslat.sh, VLLM_Baslat.sh, VRAG_Yeniden_Indeksle.sh: Linux
  karsiliklari. requirements-linux.txt ile dondurulmus surumler.
- KURULUM_UBUNTU.md: bastan sona Ubuntu kurulumu.

VLM: Ollama -> vLLM (port 8002, OpenAI-uyumlu)
- Model cyankiwi/Qwen3-VL-2B-Instruct-AWQ-4bit (4-bit AWQ).
  Kare basina ~50sn -> ~3sn.
- vlm/engine.py protokolu URL'den secer: adreste "/v1/" varsa OpenAI
  sohbet formati, yoksa Ollama /api/generate. Ayri bir "backend sec"
  ayari yok -- adresi degistirmek protokolu de degistirir.
- vLLM Ollama'dan farkli olarak VRAM yetmezse CPU'ya TASIMAZ, OOM verip
  coker. Bu yuzden VLLM_GPU_UTIL elle ayarlanir; 0.41 olculmus deger
  (0.36'da KV cache acilmiyordu).
- VLLM_KULLANIM.md: kurulum, ortam degiskenleri, VRAM butcesi, oksuz
  VLLM::EngineCore temizligi, dogrulama komutlari.

YOLO / FPS
- YOLO_QUANTIZE=16 (fp16). Ultralytics 8.4+ 'half' argumanini kaldirdi,
  yerine 'quantize' geldi.
- PROC_MAX_WIDTH=1920: kare boru hattina girmeden kuculur, SAHI patch
  sayisi alanla orantili dustugu icin 4K'da ~4x hiz.
- FULLSCAN_ALWAYS=False: her karede tam tarama kare suresini 1000ms+
  yapiyordu; artik yalniz ROI bulunamazsa fallback.
- Slicer olculen suresi ~20-25 ms/kare.

VRAG
- VRAG_DEVICE cuda:0. CPU'da worker thread'inde BLAS deadlock'una
  giriyordu (SigLIP2 Linear.forward'da donma).
- Oy-tabanli birlesik skor: benzer ucaklarda (F-16 <-> F-35,
  HURKUS <-> HURJET) tek en-yuksek hit yanlis #1 seciyordu.

LLM
- context_router.py: model -> video_id cozumu. Eski hali sabit
  video_id="live_video" kullaniyordu, tum platformlar yanlis baglamda.
- sahiplik.py: mensei (uretici ulke) ile sahiplik (isletici) ayrimi.

NOT: output/qdrant_db bilerek guncellenmedi -- yeniden indeksleme
uretilebilir ve 222 MB. veriler/ main ile birebir ayni, degisiklik yok.

+++ src/tracking/track_state.py
@@ -188,12 +188,15 @@ class Track:
         if old.vlm_class is not None:
             self.vlm_class = old.vlm_class
 
-        # BUG-FIX (KRİTİK — birleşimde sonuç kayboluyordu): vlm_done yukarıda
-        # taşınıyordu ama vlm_result/llm_result/vrag_matches HİÇ taşınmıyordu.
-        # Sonuç: okluzyon sonrası aynı ID'ye birleşen bir track, vlm_done=True
-        # (yeni VLM çağrısını engelliyor) ama vlm_result=None (LLM tetikleme
-        # kapısı "vlm_done and vlm_result" hiç açılmıyor) durumuna düşüyor —
-        # o track için bir daha asla VLM/LLM sonucu üretilemiyordu.
+        # BUG-FIX (LLM hic tetiklenmiyordu): birlestirmede vlm_done devrediliyor
+        # ama vlm_result DEVREDILMIYORDU. Sonuc kalici kilitlenme:
+        #   - LLM kapisi  `if vlm_done and vlm_result`  -> vlm_result None -> kapali
+        #   - VLM kapisi  `not vlm_done`                -> True -> kapali
+        # yani track bir kez re-ID ile birlesince vlm_result bir daha hic dolmuyor
+        # ve o track icin LLM imkansiz hale geliyordu. Ilgili LLM alanlari da
+        # (llm_result / hash / querying) tasinmali ki karar kaybolmasin ve ayni
+        # VLM sonucu icin iki kez LLM cagrilmasin. vrag_matches ulke ve baglam
+        # yonlendirmesi (context_router) icin gerekli.
         if getattr(old, "vlm_result", None) is not None and getattr(self, "vlm_result", None) is None:
             self.vlm_result = old.vlm_result
         if getattr(old, "llm_result", None) is not None and getattr(self, "llm_result", None) is None:
@@ -203,8 +206,7 @@ class Track:
         if getattr(old, "last_llm_vlm_hash", None) is not None and getattr(self, "last_llm_vlm_hash", None) is None:
             self.last_llm_vlm_hash = old.last_llm_vlm_hash
         self.is_llm_querying = getattr(self, "is_llm_querying", False) or getattr(old, "is_llm_querying", False)
-
-
+            
         self._ar_history = (old._ar_history + self._ar_history)[-15:]
         self._type_label_history = (
             getattr(old, "_type_label_history", []) + self._type_label_history
@@ -841,7 +843,7 @@ class Track:
         # BUG-FIX: Sentetik grafikler (beyaz arkaplan Ã¼stÃ¼ne siyah yazÄ±lar vb.)
         # aÅŸÄ±rÄ± yÃ¼ksek (>2000) Laplacian verir ve gerÃ§ek uÃ§aklarÄ±n Ã¶nÃ¼ne geÃ§er.
         # Laplacian (keskinlik) skoru 800.0 ile sÄ±nÄ±rlandÄ±.
-        lap_var = float(cv2.Laplacian(gray, cv2.CV_32F).var())  # CV_64F yerine CV_32F: aynı sonuç, daha az bellek/işlem
+        lap_var = float(cv2.Laplacian(gray, cv2.CV_64F).var())
         lap_var_clipped = min(800.0, lap_var)
 
         # BUG-FIX: GÃ¶kyÃ¼zÃ¼nde net bir silÃ¼et olsa bile, arkaplan Ã§ok homojen olduÄŸu iÃ§in
